style: rename file and class following operation name
