+++ src/operation/sto/transfer-by-partition.ts
@@ -8,27 +8,40 @@ import { CurrencyID } from "../../common"
 import { Big, HintedObject } from "../../types"
 import { Assert, ECODE, MitumError } from "../../error"
 
-export class RedeemTokenItem extends STOItem {
+export class TransferByPartitionItem extends STOItem {
     readonly tokenHolder: Address
-    readonly amount: Big
+    readonly receiver: Address
     readonly partition: Partition
+    readonly amount: Big
 
     constructor(
         contract: string | Address, 
         tokenHolder: string | Address,
-        amount: string | number | Big,
+        receiver: string | Address,
         partition: string | Partition,
+        amount: string | number | Big,
         currency: string | CurrencyID,
     ) {
-        super(HINT.STO.REDEEM.ITEM, contract, currency)
+        super(HINT.STO.TRANSFER_BY_PARTITION.ITEM, contract, currency)
 
         this.tokenHolder = Address.from(tokenHolder)
-        this.amount  = Big.from(amount)
+        this.receiver = Address.from(receiver)
         this.partition = Partition.from(partition)
+        this.amount  = Big.from(amount)
 
         Assert.check(
             this.contract.toString() !== this.tokenHolder.toString(),
             MitumError.detail(ECODE.INVALID_ITEM, "contract is same with token holder address")
+        )
+
+        Assert.check(
+            this.contract.toString() !== this.receiver.toString(),
+            MitumError.detail(ECODE.INVALID_ITEM, "contract is same with receiver address")
+        )
+
+        Assert.check(
+            this.tokenHolder.toString() !== this.receiver.toString(),
+            MitumError.detail(ECODE.INVALID_ITEM, "token holder is same with receiver address")
         )
 
         Assert.check(
@@ -41,8 +54,9 @@ export class RedeemTokenItem extends STOItem {
         return Buffer.concat([
             super.toBuffer(),
             this.tokenHolder.toBuffer(),
-            this.amount.toBuffer(),
+            this.receiver.toBuffer(),
             this.partition.toBuffer(),
+            this.amount.toBuffer(),
             this.currency.toBuffer(),
         ])
     }
@@ -50,28 +64,29 @@ export class RedeemTokenItem extends STOItem {
     toHintedObject(): HintedObject {
         return {
             ...super.toHintedObject(),
-            tokenHolder: this.tokenHolder.toString(),
-            amount: this.amount.toString(),
+            tokenholder: this.tokenHolder.toString(),
+            receiver: this.receiver.toString(),
             partition: this.partition.toString(),
+            amount: this.amount.toString(),
         }
     }
 
     toString(): string {
-        return this.tokenHolder.toString()
+        return `${this.tokenHolder.toString()}-${this.receiver.toString()}-${this.partition.toString()}`
     }
 }
 
-export class RedeemTokenFact extends OperationFact<RedeemTokenItem> {
-    constructor(token: string, sender: string | Address, items: RedeemTokenItem[]) {
-        super(HINT.STO.REDEEM.FACT, token, sender, items)
+export class TransferByPartitionFact extends OperationFact<TransferByPartitionItem> {
+    constructor(token: string, sender: string | Address, items: TransferByPartitionItem[]) {
+        super(HINT.STO.TRANSFER_BY_PARTITION.FACT, token, sender, items)
 
         Assert.check(
             new Set(items.map(it => it.toString())).size === items.length,
-            MitumError.detail(ECODE.INVALID_ITEMS, "duplicate token holder found in items")
+            MitumError.detail(ECODE.INVALID_ITEMS, "duplicate token holder-receiver-partition found in items")
         )
     }
 
     get operationHint() {
-        return HINT.STO.REDEEM.OPERATION
+        return HINT.STO.TRANSFER_BY_PARTITION.OPERATION
     }
 }
